[DP-76] fix: 데이터양 타입 수정 #58

+++ src/main/java/com/dapanda/plan/service/PlanService.java
@@ -31,7 +31,7 @@ public class PlanService {
 		PlanCategory category = randomCategory();
 		AgeGroup ageGroup = randomAgeGroup();
 		String name = randomPlanName(category, ageGroup);
-		int dataAmount = randomDataAmount();
+		float dataAmount = randomDataAmount();
 		int price = randomMonthlyPrice();
 
 		Plan plan = Plan.of(
